requirements.txt
fastapi
uvicorn
streamlit
pydantic
gitpython
openai